refactor: api 문서 생성용 annotation 패키지 구조 수정
