+++ src/main/java/roomescape/controller/apidocs/MemberApi.java
@@ -11,7 +11,7 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import roomescape.annotation.SecurityDocs;
+import roomescape.controller.apidocs.annotation.SecurityDocs;
 import roomescape.dto.request.SignupRequest;
 import roomescape.dto.response.MemberResponse;
 import roomescape.exception.dto.ErrorResponse;
