+++ src/main/java/roomescape/controller/apidocs/ThemeApi.java
@@ -11,7 +11,7 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import roomescape.annotation.SecurityDocs;
+import roomescape.controller.apidocs.annotation.SecurityDocs;
 import roomescape.dto.request.CreateThemeRequest;
 import roomescape.dto.response.ThemeResponse;
 import roomescape.exception.dto.ErrorResponse;
